refactor: split TestDispatchFoodTruckHaltEnforcement into test_tools_dispatch_halt.py
Extract halt enforcement tests (7 test methods + _write_campaign_state helper)
from test_tools_dispatch.py into a new test_tools_dispatch_halt.py. Move shared
helpers _make_recipe_info and _make_standard_recipe to tests/server/_helpers.py.
Update _test_filter.py cascade map and server CLAUDE.md.
