+++ tests/server/_helpers.py
@@ -4,6 +4,19 @@ from __future__ import annotations
 
 from autoskillit.core import SkillResult
 from autoskillit.core.types import RetryReason
+from tests.fleet._helpers import _make_recipe_info as _fleet_make_recipe_info
+
+
+def _make_recipe_info(name: str = "test-recipe"):
+    return _fleet_make_recipe_info(name, path_prefix="/fake/recipes/")
+
+
+def _make_standard_recipe(name: str = "test-recipe", ingredient_keys: list[str] | None = None):
+    """Return a minimal Recipe with kind=STANDARD for use as load_recipe mock return value."""
+    from autoskillit.recipe.schema import Recipe, RecipeIngredient, RecipeKind
+
+    ingredients = {k: RecipeIngredient(description=k) for k in (ingredient_keys or [])}
+    return Recipe(name=name, description="test", ingredients=ingredients, kind=RecipeKind.STANDARD)
 
 
 def _skill_ok(report_text: str = "## Bug Report\ndetails") -> SkillResult:
